pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "serverlens"
version = "2.0.0"
description = "ServerLens — Secure Read-Only Server Diagnostics MCP Server"
license = {text = "Proprietary"}
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
    "argon2-cffi>=23.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
serverlens = "serverlens.__main__:main"

[tool.setuptools.packages.find]
include = ["serverlens*", "serverlens_mcp*"]

[tool.pytest.ini_options]
testpaths = ["tests_py"]
asyncio_mode = "auto"
